reorganize misisng coverage test files
